[master] Fix a type error from size_t to double
[ 97%] Building CXX object simulator/unit_tests/CMakeFiles/simulator_tests.dir/src/IntegrationTests.cpp.obj
d:/GJ/GIT/SimulateurIrtJv/code/simulator/unit_tests/src/IntegrationTests.cpp: In member function 'virtual void IntegrationTests_can_simulate_oscillating_cube_Test::TestBody()':
d:/GJ/GIT/SimulateurIrtJv/code/simulator/unit_tests/src/IntegrationTests.cpp:116:28: error: conversion to 'double' from 'std::size_t {aka long long unsigned int}' may alter its value [-Werror=conversion]
         const double t = i*dt;
                            ^
compilation terminated due to -Wfatal-errors.
cc1plus.exe: all warnings being treated as errors
make[2]: *** [simulator/unit_tests/CMakeFiles/simulator_tests.dir/src/IntegrationTests.cpp.obj] Error 1
make[1]: *** [simulator/unit_tests/CMakeFiles/simulator_tests.dir/all] Error 2
make: *** [all] Error 2

+++ code/simulator/unit_tests/src/IntegrationTests.cpp
@@ -113,7 +113,7 @@ TEST_F(IntegrationTests, can_simulate_oscillating_cube)
     const double z0 = L/2;
     for (size_t i = 0 ; i < N ; ++i)
     {
-        const double t = i*dt;
+        const double t = double(i)*dt;
         ASSERT_EQ(1+13, res.at(i).size())          << "Time step: i=" << i;
         ASSERT_DOUBLE_EQ(t, res.at(i)["t"])        << "Time step: i=" << i;
         ASSERT_NEAR(0, res.at(i)["x(cube)"], eps)  << "Time step: i=" << i;
